+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - Added Email validator
+
+* Added a validator `isEmail` in `AutoValidators`
+* Removed red fill from text fields
+
 ## [0.0.1] - Initial sketch of the package
 
 * Current this package only supports creating text fields.